+++ tests/dummy/routes/json-api.php
@@ -20,29 +20,28 @@ use CloudCreativity\LaravelJsonApi\Routing\ApiGroup;
 
 JsonApi::register('default', [], function (ApiGroup $api) {
     $api->resource('comments', [
+        'controller' => true,
         'has-one' => 'commentable',
     ]);
     $api->resource('countries', [
-        'controller' => false,
         'has-many' => ['users', 'posts'],
     ]);
     $api->resource('posts', [
+        'controller' => true,
         'has-one' => [
             'author' => ['inverse' => 'users']
         ],
         'has-many' => ['comments', 'tags'],
     ]);
     $api->resource('users', [
-        'controller' => false,
         'has-one' => 'phone',
     ]);
-    $api->resource('videos', [
-        'controller' => false,
-    ]);
+    $api->resource('videos');
     $api->resource('tags', [
-        'controller' => false,
         'has-many' => 'taggables',
     ]);
 
-    $api->resource('sites');
+    $api->resource('sites', [
+        'controller' => true,
+    ]);
 });