+++ CHANGELOG.md
@@ -16,13 +16,13 @@ in the `ErrorBag` class.
 - JSON API resource paths are now automatically converted to model relationship paths for eager loading in
 the Eloquent adapter.
 - The Eloquent adapter now applies eager loading when reading or updating a specific resource.
-- Resources no longer need to have a controller as the generic JSON API controller will now handle any resource.
-Any resource that does not have a controller must use `false` as its `controller` option when registering the 
-resource routes.
 - Eloquent adapters can now *guard* JSON API fields via their `$guarded` attribute. Guarded fields will be skipped
 when filling attributes into the model.
 
 ### Changed
+- By default resources no longer need to have a controller as the generic JSON API controller will now
+handle any resource. If resources have controllers, the `controller` routing option can be set to a string
+controller name, or `true` to use a controller with the same name as the resource.
 - Split adapter into resource and relationship adapter, and created classes to specifically deal with Eloquent
 relationships.
 - Adapters now handle both reading and modifying domain records.